Update xterm to version 374 ok tb@

+++ app/xterm/trace.h
@@ -1,4 +1,4 @@
-/* $XTermId: trace.h,v 1.93 2022/02/21 22:54:05 tom Exp $ */
+/* $XTermId: trace.h,v 1.94 2022/10/06 20:40:49 tom Exp $ */
 
 /*
  * Copyright 1997-2021,2022 by Thomas E. Dickey
@@ -69,9 +69,9 @@ extern	void	TraceClose (void);
 #define TRACE_L "{{"
 #define TRACE_R "}}"
 
-extern	const char * visibleChars(const Char * /* buf */, unsigned /* len */);
+extern	const char * visibleChars(const Char * /* buf */, size_t /* len */);
 extern	const char * visibleEventMode(EventMode);
-extern	const char * visibleIChars(const IChar * /* buf */, unsigned /* len */);
+extern	const char * visibleIChars(const IChar * /* buf */, size_t /* len */);
 extern	const char * visibleUChar(unsigned);
 extern	const char * visibleDblChrset(unsigned /* chrset */);
 extern	const char * visibleEventType (int);
